Create basic query history test
