+++ public/app/features/explore/RichHistory/RichHistoryContainer.tsx
@@ -58,15 +58,17 @@ export function RichHistoryContainer(props: Props) {
         setHeight(Number(ref.style.height.slice(0, -2)));
       }}
     >
-      <RichHistory
-        richHistory={richHistory}
-        firstTab={firstTab}
-        activeDatasourceInstance={activeDatasourceInstance}
-        exploreId={exploreId}
-        deleteRichHistory={deleteRichHistory}
-        onClose={onClose}
-        height={height}
-      />
+      <div data-testid="richHistory">
+        <RichHistory
+          richHistory={richHistory}
+          firstTab={firstTab}
+          activeDatasourceInstance={activeDatasourceInstance}
+          exploreId={exploreId}
+          deleteRichHistory={deleteRichHistory}
+          onClose={onClose}
+          height={height}
+        />
+      </div>
     </ExploreDrawer>
   );
 }
