+++ public/app/features/explore/Wrapper.test.tsx
@@ -1,32 +1,12 @@
 import React from 'react';
-import { fireEvent, render, screen, waitFor } from '@testing-library/react';
-import Wrapper from './Wrapper';
-import { configureStore } from '../../store/configureStore';
-import { Provider } from 'react-redux';
-import { locationService, setDataSourceSrv, setEchoSrv } from '@grafana/runtime';
-import {
-  ArrayDataFrame,
-  DataQueryResponse,
-  DataSourceApi,
-  DataSourceInstanceSettings,
-  FieldType,
-  QueryEditorProps,
-  ScopedVars,
-  serializeStateToUrlParam,
-} from '@grafana/data';
-import { selectors } from '@grafana/e2e-selectors';
-
-import { from, Observable } from 'rxjs';
-import { LokiDatasource } from '../../plugins/datasource/loki/datasource';
-import { LokiQuery } from '../../plugins/datasource/loki/types';
-import { fromPairs } from 'lodash';
-import { EnhancedStore } from '@reduxjs/toolkit';
+import { fireEvent, screen, waitFor } from '@testing-library/react';
+import { locationService } from '@grafana/runtime';
+import { serializeStateToUrlParam } from '@grafana/data';
 import userEvent from '@testing-library/user-event';
 import { splitOpen } from './state/main';
-import { Route, Router } from 'react-router-dom';
-import { GrafanaRoute } from 'app/core/navigation/GrafanaRoute';
-import { initialUserState } from '../profile/state/reducers';
-import { Echo } from 'app/core/services/echo/Echo';
+import { setupExplore } from './spec/helper/exploreSetup';
+import { makeLogsQueryResponse, makeMetricsQueryResponse } from './spec/helper/query';
+import { changeDatasource } from './spec/helper/interactions';
 
 type Mock = jest.Mock;
 
@@ -53,13 +33,13 @@ jest.mock('react-virtualized-auto-sizer', () => {
 
 describe('Wrapper', () => {
   it('shows warning if there are no data sources', async () => {
-    setup({ datasources: [] });
+    setupExplore({ datasources: [] });
     // Will throw if isn't found
     screen.getByText(/Explore requires at least one data source/i);
   });
 
   it('inits url and renders editor but does not call query on empty url', async () => {
-    const { datasources } = setup();
+    const { datasources } = setupExplore();
 
     // Wait for rendering the editor
     await screen.findByText(/Editor/i);
@@ -84,7 +64,7 @@ describe('Wrapper', () => {
         range: { from: 'now-1h', to: 'now' },
       }),
     };
-    const { datasources, store } = setup({ query });
+    const { datasources, store } = setupExplore({ query });
     (datasources.loki.query as Mock).mockReturnValueOnce(makeLogsQueryResponse());
 
     // Make sure we render the logs panel
@@ -116,7 +96,7 @@ describe('Wrapper', () => {
 
   it('handles url change and runs the new query', async () => {
     const query = { left: JSON.stringify(['now-1h', 'now', 'loki', { expr: '{ label="value"}' }]) };
-    const { datasources } = setup({ query });
+    const { datasources } = setupExplore({ query });
     (datasources.loki.query as Mock).mockReturnValueOnce(makeLogsQueryResponse());
     // Wait for rendering the logs
     await screen.findByText(/custom log line/i);
@@ -135,7 +115,7 @@ describe('Wrapper', () => {
 
   it('handles url change and runs the new query with different datasource', async () => {
     const query = { left: JSON.stringify(['now-1h', 'now', 'loki', { expr: '{ label="value"}' }]) };
-    const { datasources } = setup({ query });
+    const { datasources } = setupExplore({ query });
     (datasources.loki.query as Mock).mockReturnValueOnce(makeLogsQueryResponse());
     // Wait for rendering the logs
     await screen.findByText(/custom log line/i);
@@ -155,7 +135,7 @@ describe('Wrapper', () => {
 
   it('handles changing the datasource manually', async () => {
     const query = { left: JSON.stringify(['now-1h', 'now', 'loki', { expr: '{ label="value"}', refId: 'A' }]) };
-    const { datasources } = setup({ query });
+    const { datasources } = setupExplore({ query });
     (datasources.loki.query as Mock).mockReturnValueOnce(makeLogsQueryResponse());
     // Wait for rendering the editor
     await screen.findByText(/Editor/i);
@@ -174,7 +154,7 @@ describe('Wrapper', () => {
   });
 
   it('opens the split pane when split button is clicked', async () => {
-    setup();
+    setupExplore();
     // Wait for rendering the editor
     const splitButton = await screen.findByText(/split/i);
     fireEvent.click(splitButton);
@@ -198,7 +178,7 @@ describe('Wrapper', () => {
       }),
     };
 
-    const { datasources } = setup({ query });
+    const { datasources } = setupExplore({ query });
     (datasources.loki.query as Mock).mockReturnValueOnce(makeLogsQueryResponse());
     (datasources.elastic.query as Mock).mockReturnValueOnce(makeLogsQueryResponse());
 
@@ -239,7 +219,7 @@ describe('Wrapper', () => {
       left: JSON.stringify(['now-1h', 'now', 'loki', { refId: 'A' }]),
       right: JSON.stringify(['now-1h', 'now', 'elastic', { refId: 'A' }]),
     };
-    setup({ query });
+    setupExplore({ query });
     const closeButtons = await screen.findAllByTitle(/Close split pane/i);
     userEvent.click(closeButtons[1]);
 
@@ -253,7 +233,7 @@ describe('Wrapper', () => {
     const query = {
       left: JSON.stringify(['now-1h', 'now', 'loki', { expr: '{ label="value"}' }]),
     };
-    const { datasources } = setup({ query });
+    const { datasources } = setupExplore({ query });
     (datasources.loki.query as Mock).mockReturnValue(makeLogsQueryResponse());
     (datasources.elastic.query as Mock).mockReturnValue(makeLogsQueryResponse());
 
@@ -271,7 +251,7 @@ describe('Wrapper', () => {
     const query = {
       left: JSON.stringify(['now-1h', 'now', 'loki', { expr: '{ label="value"}' }]),
     };
-    const { datasources, store } = setup({ query });
+    const { datasources, store } = setupExplore({ query });
     (datasources.loki.query as Mock).mockReturnValue(makeLogsQueryResponse());
     (datasources.elastic.query as Mock).mockReturnValue(makeLogsQueryResponse());
 
@@ -290,7 +270,7 @@ describe('Wrapper', () => {
     const query = {
       left: JSON.stringify(['now-1h', 'now', 'loki', { expr: '{ label="value"}' }]),
     };
-    const { datasources } = setup({ query });
+    const { datasources } = setupExplore({ query });
     (datasources.loki.query as Mock).mockReturnValue(makeLogsQueryResponse());
     // This is mainly to wait for render so that the left pane state is initialized as that is needed for the title
     // to include the datasource
@@ -302,7 +282,7 @@ describe('Wrapper', () => {
     const query = {
       left: JSON.stringify(['now-1h', 'now', 'loki', { expr: '{ label="value"}' }]),
     };
-    const { datasources, store } = setup({ query });
+    const { datasources, store } = setupExplore({ query });
     (datasources.loki.query as Mock).mockReturnValue(makeLogsQueryResponse());
     (datasources.elastic.query as Mock).mockReturnValue(makeLogsQueryResponse());
 
@@ -315,141 +295,9 @@ describe('Wrapper', () => {
   });
 
   it('removes `from` and `to` parameters from url when first mounted', () => {
-    setup({ searchParams: 'from=1&to=2&orgId=1' });
+    setupExplore({ searchParams: 'from=1&to=2&orgId=1' });
 
     expect(locationService.getSearchObject()).toEqual(expect.not.objectContaining({ from: '1', to: '2' }));
     expect(locationService.getSearchObject()).toEqual(expect.objectContaining({ orgId: '1' }));
   });
 });
-
-type DatasourceSetup = { settings: DataSourceInstanceSettings; api: DataSourceApi };
-type SetupOptions = {
-  datasources?: DatasourceSetup[];
-  query?: any;
-  searchParams?: string;
-};
-
-function setup(options?: SetupOptions): { datasources: { [name: string]: DataSourceApi }; store: EnhancedStore } {
-  // Clear this up otherwise it persists data source selection
-  // TODO: probably add test for that too
-  window.localStorage.clear();
-
-  // Create this here so any mocks are recreated on setup and don't retain state
-  const defaultDatasources: DatasourceSetup[] = [
-    makeDatasourceSetup(),
-    makeDatasourceSetup({ name: 'elastic', id: 2 }),
-  ];
-
-  const dsSettings = options?.datasources || defaultDatasources;
-
-  setDataSourceSrv({
-    getList(): DataSourceInstanceSettings[] {
-      return dsSettings.map((d) => d.settings);
-    },
-    getInstanceSettings(name: string) {
-      return dsSettings.map((d) => d.settings).find((x) => x.name === name || x.uid === name);
-    },
-    get(name?: string | null, scopedVars?: ScopedVars): Promise<DataSourceApi> {
-      return Promise.resolve(
-        (name ? dsSettings.find((d) => d.api.name === name || d.api.uid === name) : dsSettings[0])!.api
-      );
-    },
-  } as any);
-
-  setEchoSrv(new Echo());
-
-  const store = configureStore();
-  store.getState().user = {
-    ...initialUserState,
-    orgId: 1,
-    timeZone: 'utc',
-  };
-
-  store.getState().navIndex = {
-    explore: {
-      id: 'explore',
-      text: 'Explore',
-      subTitle: 'Explore your data',
-      icon: 'compass',
-      url: '/explore',
-    },
-  };
-
-  locationService.push({ pathname: '/explore', search: options?.searchParams });
-
-  if (options?.query) {
-    locationService.partial(options.query);
-  }
-
-  const route = { component: Wrapper };
-
-  render(
-    <Provider store={store}>
-      <Router history={locationService.getHistory()}>
-        <Route path="/explore" exact render={(props) => <GrafanaRoute {...props} route={route as any} />} />
-      </Router>
-    </Provider>
-  );
-
-  return { datasources: fromPairs(dsSettings.map((d) => [d.api.name, d.api])), store };
-}
-
-function makeDatasourceSetup({ name = 'loki', id = 1 }: { name?: string; id?: number } = {}): DatasourceSetup {
-  const meta: any = {
-    info: {
-      logos: {
-        small: '',
-      },
-    },
-    id: id.toString(),
-  };
-  return {
-    settings: {
-      id,
-      uid: name,
-      type: 'logs',
-      name,
-      meta,
-      access: 'proxy',
-      jsonData: {},
-    },
-    api: {
-      components: {
-        QueryEditor(props: QueryEditorProps<LokiDatasource, LokiQuery>) {
-          return (
-            <div>
-              {name} Editor input: {props.query.expr}
-            </div>
-          );
-        },
-      },
-      name: name,
-      uid: name,
-      query: jest.fn(),
-      getRef: jest.fn(),
-      meta,
-    } as any,
-  };
-}
-
-function makeLogsQueryResponse(marker = ''): Observable<DataQueryResponse> {
-  const df = new ArrayDataFrame([{ ts: Date.now(), line: `custom log line ${marker}` }]);
-  df.meta = {
-    preferredVisualisationType: 'logs',
-  };
-  df.fields[0].type = FieldType.time;
-  return from([{ data: [df] }]);
-}
-
-function makeMetricsQueryResponse(): Observable<DataQueryResponse> {
-  const df = new ArrayDataFrame([{ ts: Date.now(), val: 1 }]);
-  df.fields[0].type = FieldType.time;
-  return from([{ data: [df] }]);
-}
-
-async function changeDatasource(name: string) {
-  const datasourcePicker = (await screen.findByLabelText(selectors.components.DataSourcePicker.container)).children[0];
-  fireEvent.keyDown(datasourcePicker, { keyCode: 40 });
-  const option = screen.getByText(name);
-  fireEvent.click(option);
-}
